сделал 7. Личный проект: время разбивать камни
